% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andersen.gill.score.r
\name{andersen.gill.score}
\alias{andersen.gill.score}
\title{GWAS with Andersen-Gill score test}
\usage{
andersen.gill.score(x, null.model, beg = 1, end = ncol(x))
}
\arguments{
\item{x}{a bedmatrix}

\item{null.model}{an object of class nullModel}

\item{beg}{Index of the first SNP tested for association}

\item{end}{Index of the last SNP tested for association}
}
\value{
A data frame giving for each SNP the association statistics.
}
\description{
GWAS with Andersen-Gill score test
}
\examples{
# TBC

}
