% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.test.r
\name{score.test}
\alias{score.test}
\title{Score test for Andersen-Gill model}
\usage{
score.test(null.model, X)
}
\arguments{
\item{null.model}{an object of class nullModel}

\item{X}{a vector}
}
\description{
Score test for Andersen-Gill model
}
