% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scottage.r
\name{scottage}
\alias{scottage}
\title{Scottage options}
\usage{
scottage(...)
}
\arguments{
\item{...}{options to be defined, using 'name = value', or name(s) of 
option(s) to get.}
}
\value{
A list with the defined options, or a single element when \code{scottage(name)}
is used.
}
\description{
Set or get options values for package 'scottage'
}
\details{
Currently, only one option can be defined, \code{use.float}
The default value is set to 'FALSE'
}
\examples{
old.value <- scottage("use.float")
scottage(use.float = TRUE)
scottage("use.float")
scottage(use.float = old.value)
}
